// frontend/src/api/models/listFinanceReceivableReceivablesParams.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { ListFinanceReceivableReceivablesSourceType } from './listFinanceReceivableReceivablesSourceType'
import type { ListFinanceReceivableReceivablesStatus } from './listFinanceReceivableReceivablesStatus'

export type ListFinanceReceivableReceivablesParams = {
  /**
   * Search term (receivable number, customer name)
   */
  search?: string
  /**
   * Customer ID
   */
  customer_id?: string
  /**
   * Status
   */
  status?: ListFinanceReceivableReceivablesStatus
  /**
   * Source type
   */
  source_type?: ListFinanceReceivableReceivablesSourceType
  /**
   * From date (ISO 8601)
   */
  from_date?: string
  /**
   * To date (ISO 8601)
   */
  to_date?: string
  /**
   * Filter overdue only
   */
  overdue?: boolean
  /**
   * Page number
   */
  page?: number
  /**
   * Page size
   * @maximum 100
   */
  page_size?: number
}