// frontend/src/api/models/listFinanceReceivableReceivablesStatus.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type ListFinanceReceivableReceivablesStatus =
  (typeof ListFinanceReceivableReceivablesStatus)[keyof typeof ListFinanceReceivableReceivablesStatus]

export const ListFinanceReceivableReceivablesStatus = {
  PENDING: 'PENDING',
  PARTIAL: 'PARTIAL',
  PAID: 'PAID',
  REVERSED: 'REVERSED',
  CANCELLED: 'CANCELLED',
} as const