// frontend/src/api/models/listFinanceReceivableReceivablesSourceType.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type ListFinanceReceivableReceivablesSourceType =
  (typeof ListFinanceReceivableReceivablesSourceType)[keyof typeof ListFinanceReceivableReceivablesSourceType]

export const ListFinanceReceivableReceivablesSourceType = {
  SALES_ORDER: 'SALES_ORDER',
  SALES_RETURN: 'SALES_RETURN',
  MANUAL: 'MANUAL',
} as const