// frontend/src/api/models/handlerReceivableSummaryResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

/**
 * Receivable summary response
 */
export interface HandlerReceivableSummaryResponse {
  overdue_count?: number
  partial_count?: number
  pending_count?: number
  total_outstanding?: number
  total_overdue?: number
}